Sync study data between legacy designer & new study controller
